renamed application and package name
