+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "BeaconAdventure",
-  "displayName": "BeaconAdventure"
+  "name": "BeaconSuedtirolAltoAdigeAdventure",
+  "displayName": "BeaconSuedtirolAltoAdigeAdventure"
 }
