+++ ios/BeaconSuedtirolAltoAdigeAdventure-tvOSTests/BeaconAdventure-tvOSTests/Info.plist
