+++ ios/BeaconSuedtirolAltoAdigeAdventure.xcodeproj/BeaconAdventure.xcodeproj/xcshareddata/xcschemes/BeaconAdventure-tvOS.xcscheme
