+++ ios/BeaconSuedtirolAltoAdigeAdventure.xcodeproj/BeaconAdventure.xcodeproj/xcshareddata/xcschemes/BeaconAdventure.xcscheme
