+++ ios/BeaconSuedtirolAltoAdigeAdventure.xcworkspace/BeaconAdventure.xcworkspace/contents.xcworkspacedata
