+++ ios/BeaconSuedtirolAltoAdigeAdventure/BeaconAdventure/AppDelegate.h
