+++ ios/BeaconSuedtirolAltoAdigeAdventure/BeaconAdventure/AppDelegate.m
