+++ ios/BeaconSuedtirolAltoAdigeAdventure/BeaconAdventure/Base.lproj/LaunchScreen.xib
