+++ ios/BeaconSuedtirolAltoAdigeAdventure/BeaconAdventure/Images.xcassets/AppIcon.appiconset/Contents.json
