+++ ios/BeaconSuedtirolAltoAdigeAdventure/BeaconAdventure/Images.xcassets/Contents.json
