+++ ios/BeaconSuedtirolAltoAdigeAdventure/BeaconAdventure/Info.plist
