+++ ios/BeaconSuedtirolAltoAdigeAdventure/BeaconAdventure/main.m
