+++ ios/BeaconSuedtirolAltoAdigeAdventureTests/BeaconAdventureTests/BeaconAdventureTests.m
