+++ ios/BeaconSuedtirolAltoAdigeAdventureTests/BeaconAdventureTests/Info.plist
